refactor: move error handler to utils file
