+++ packages/portal/src/server-middleware/api/index.js
@@ -58,16 +58,4 @@ app.get('/version', version);
 
 app.all('/*', (req, res) => res.sendStatus(404));
 
-export const errorHandler = (res, error) => {
-  let status = error.status || 500;
-  let message = error.message;
-
-  if (error.response) {
-    status = error.response.status;
-    message = error.response.data.errorMessage;
-  }
-
-  res.status(status).set('Content-Type', 'text/plain').send(message);
-};
-
 export default app;
