+++ packages/portal/src/server-middleware/api/utils.js
@@ -1,0 +1,11 @@
+export const errorHandler = (res, error) => {
+  let status = error.status || 500;
+  let message = error.message;
+
+  if (error.response) {
+    status = error.response.status;
+    message = error.response.data.errorMessage;
+  }
+
+  res.status(status).set('Content-Type', 'text/plain').send(message);
+};
